Added onCustomRender; added in ability to edit selected row via the edit panel with prefilled data; added ability for new panel edit (single/bulk) to have custom columns
onCustomRender allows rendering of items without affecting "editing" capability.

enablePanelEdit will enabled grid to render a new icon to allow user to edit a selected row item in the edit panel. Row data will be pre-filled in the panel. This is different from the "bulk edit" which is enabled on one or more row selections and which doesn't prefill data.

customEditPanelColumns allows for customised panel edit columns for both single/bulk panel edits. This allows the grid to display a subset of the data and allow for different data to be editable via the edit panels.