User experience enhancements
Marquee Selection now optional in EditableGrid options. This is a way to get rid of the current "mouse drag" selection which doesn't select multiple rows anyways.

"Filter" modal now provides a better UX by showing/disabling dropdowns based on state.

Provided ability for columns to have the "isSortedByDefault" prop. It allows default column sorting when the grid is rendered.

Provided prependRowEditActions prop on EditableGrid when enableRowEdit is set to TRUE. This allows the "Actions" column to appear as the first column on the grid (rather than the last).